+++ server/src/server.ts
@@ -7,13 +7,17 @@ import {
     TextDocumentSyncKind,
     InitializeResult,
     DocumentLinkParams,
-    DocumentLink
+    Location,
+    Range,
 } from 'vscode-languageserver/node';
 import {
+    Position,
     TextDocument
 } from 'vscode-languageserver-textdocument';
-import OMTLinkProvider from './omtLinkProvider';
+import OmtDocumentInformationProvider from './omtDocumentInformationProvider';
 import { WorkspaceLookup } from './workspaceLookup';
+import * as fs from "fs";
+import { OmtAvailableObjects, OmtDocumentInformation, OmtLocalObject } from './types';
 
 // Create a connection for the server, using Node's IPC as a transport.
 // Also include all preview / proposed LSP features.
@@ -21,11 +25,12 @@ const connection = createConnection(ProposedFeatures.all);
 
 // Create a simple text document manager.
 const documents = new TextDocuments(TextDocument);
+const documentResults: Map<string, OmtDocumentInformation> = new Map();
 
 let hasConfigurationCapability = false;
 let hasWorkspaceFolderCapability = false;
 let hasDocumentLinkCapabilities = false;
-let omtLinkProvider: OMTLinkProvider;
+let omtDocumentInformationProvider: OmtDocumentInformationProvider;
 let workspaceLookup: WorkspaceLookup;
 
 // tell the client what functionality is supported in this LSP when it initializes.
@@ -41,7 +46,8 @@ connection.onInitialize((params: InitializeParams) => {
     const result: InitializeResult = {
         capabilities: {
             textDocumentSync: TextDocumentSyncKind.Incremental,
-            documentLinkProvider: {resolveProvider: true},
+            definitionProvider: true,
+            documentLinkProvider: { resolveProvider: false },
         }
     };
     if (hasWorkspaceFolderCapability) {
@@ -58,7 +64,7 @@ connection.onInitialize((params: InitializeParams) => {
 connection.onInitialized(() => {
     shutdownCheck();
     workspaceLookup = new WorkspaceLookup(connection.workspace);
-    omtLinkProvider = new OMTLinkProvider(workspaceLookup);
+    omtDocumentInformationProvider = new OmtDocumentInformationProvider(workspaceLookup);
     workspaceLookup.init().then(() => {
         workspaceLookup.scanAll();
     });
@@ -69,9 +75,59 @@ connection.onInitialized(() => {
     }
     if (hasDocumentLinkCapabilities) {
         connection.onDocumentLinks(documentLinksHandler);
-        connection.onDocumentLinkResolve(documentLinkResolve);
     }
 });
+
+function comparePositions(pos1: Position, pos2: Position) {
+    if (pos1.line !== pos2.line) {
+        return pos1.line > pos2.line ? 1 : -1;
+    }
+    if (pos1.character !== pos2.character) {
+        return pos1.character > pos2.character ? 1 : -1;
+    }
+    return 0;
+}
+
+function positionInRange(position: Position, range: Range) {
+    return comparePositions(range.start, position) <= 0 && comparePositions(position, range.end) <= 0;
+}
+
+connection.onDefinition((params) => {
+    const locations: Location[] = [];
+    const document = documents.get(params.textDocument.uri);
+    if (document) {
+        const availableObjects = getOmtDocumentResult(document);
+
+        availableObjects.calledObjects.forEach(link => {
+            if (positionInRange(params.position, link.range)) {
+                locations.push(...getLocationsForLink(params.textDocument.uri, availableObjects, link));
+            }
+        });
+    }
+
+    return locations;
+});
+
+function getLocationsForLink(linkUrl: string, availableObjects: OmtAvailableObjects, link: OmtLocalObject): Location[] {
+    const locations: Location[] = []
+    availableObjects.definedObjects.filter(x => x.name === link.name).forEach(t => {
+        locations.push(Location.create(linkUrl, t.range));
+    });
+    availableObjects.availableImports.filter(x => x.name === link.name).forEach(i => {
+        const newLinkUrl = `${i.fullUrl}`;
+        locations.push(...getLocationsInImportedFile(newLinkUrl, link));
+    });
+    return locations;
+}
+
+function getLocationsInImportedFile(linkUrl: string, link: OmtLocalObject): Location[] {
+    const otherDocument = TextDocument.create(linkUrl, 'omt', 1, fs.readFileSync(linkUrl).toString());
+    if (otherDocument) {
+        const availableObjects = omtDocumentInformationProvider.provideAvailableObjectsFromDocument(otherDocument);
+        return getLocationsForLink(linkUrl, availableObjects, link);
+    }
+    return [];
+}
 
 /**
  * shutdown protocol
@@ -103,12 +159,29 @@ connection.onExit(() => {
     process.exit();
 });
 
+let timerId: NodeJS.Timeout;
+let currentChangingDocumentUri: string;
+
 // The content of a text document has changed. This event is emitted
 // when the text document first opened or when its content has changed.
 documents.onDidChangeContent((change) => {
     shutdownCheck();
+    if (currentChangingDocumentUri === change.document.uri) {
+        timerId && clearTimeout(timerId);
+    }
+    timerId = setTimeout(() => documentResults.set(change.document.uri, omtDocumentInformationProvider.provideDocumentInformation(change.document)), 1000);
+    currentChangingDocumentUri = change.document.uri;
     return workspaceLookup.fileChanged(change);
 });
+
+function getOmtDocumentResult(document: TextDocument): OmtDocumentInformation {
+    let result = documentResults.get(document.uri);
+    if (!result) {
+        result = omtDocumentInformationProvider.provideDocumentInformation(document);
+        documentResults.set(document.uri, result);
+    }
+    return result;
+}
 
 /**
  * find all document links in the document.
@@ -122,23 +195,10 @@ const documentLinksHandler = (params: DocumentLinkParams) => {
     shutdownCheck();
     const document = documents.get(params.textDocument.uri);
     if (document) {
-        return omtLinkProvider.provideDocumentLinks(document)
+        return getOmtDocumentResult(document).documentLinks;
     } else {
         return undefined;
     }
-}
-
-/**
- * called by the client when a link target was empty.
- * it will try and resolve it using the data on the link.
- * @param link a documenet link without a target
- */
-const documentLinkResolve = (link: DocumentLink) => {
-    shutdownCheck();
-    // the data would have been set during a call to `documentLinksHandler` when the document was opened
-    // usually because it would be less efficient to resolve the link at that time.
-    link.target = omtLinkProvider.resolveLink(link.data);
-    return link;
 }
 
 // Make the text document manager listen on the connection